Refactor cobra stage command
Renamed it and tidied it up a bit
